% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Description of a data object}
\usage{
describe()
}
\description{
Description of a data object like a data.frame that is more comprehensive than basic R functions
like str() or summary()
}
