% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hi}
\alias{hi}
\title{Show package documentation index (help index)}
\usage{
hi(package)
}
\arguments{
\item{package}{name of the package}
}
\description{
This is just a shorter way of calling \code{help(package = "package name")}.
If you use RStudio you do not have to type the quotation marks (""),
just autocomplete the package name by pressing TAB.
}
\examples{
hi("bodomisc")
help(package = "bodomisc")

}
