% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_essential_packages.R
\name{install_essential_packages}
\alias{install_essential_packages}
\title{https://stackoverflow.com/questions/4090169/elegant-way-to-check-for-missing-packages-and-install-them}
\usage{
install_essential_packages()
}
\description{
https://stackoverflow.com/questions/4090169/elegant-way-to-check-for-missing-packages-and-install-them
}
