% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{majority}
\alias{majority}
\title{Return most frequent level of factor (majority voting)}
\usage{
majority(x)
}
\arguments{
\item{x}{a factor}
}
\value{
factor of length 1 with the same levels as x
}
\description{
Return most frequent level of factor (majority voting)
}
\section{TODO}{

  If the majority voting is ambiguous: random choice or preferred level.
}

\examples{
x1 = c(2,3,1,1,1,3,3,2,2,1)
f1 = factor(x1, levels = c(1, 2, 3), labels = c("red", "blue", "green"))
f1
majority(f1)

x2 = c(2,3,1,1,1,3,3,2,2,1,2)
f2 = factor(x2, levels = c(1, 2, 3), labels = c("red", "blue", "green"))
f2
majority(f2)

x3 = c(3,3,3,1,1,1,1,5,5,3,1)
o3 = ordered(x2, levels = c(1, 3, 5), labels = c("Low", "Medium", "High"))
o3
majority(o3)
}
