% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{normalize}
\alias{normalize}
\title{Scale data vector to 0-1-range (or another selectable range)}
\usage{
normalize(x, lower.bound = 0, upper.bound = 1)
}
\arguments{
\item{x}{numeric}

\item{lower.bound=0}{Lowest value after scaling.}

\item{upper.bound=1}{Highest value after scaling.}
}
\description{
Alternative names: min-max-scaling, data normalization.
}
\details{
For standardazing with sample mean and sample standard deviation use base::scale().

See \href{https://en.wikipedia.org/wiki/Feature_scaling}{Feature Scaling Wikipedia},
\href{https://en.wikipedia.org/wiki/Normalization_(statistics)}{Normalization Wikipedia}.

TODO: support matrices and data.frames
}
\examples{
normalize(1:101)
normalize(rnorm(10, -10, 10), lower.bound = 0, upper.bound = 100)

}
