% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotClassification2D}
\alias{plotClassification2D}
\title{Visualize a classification task in 2D}
\usage{
plotClassification2D(task, model, grid.res = 100, x1.lim = NULL,
  x2.lim = NULL, colours = FALSE)
}
\arguments{
\item{task}{mlr classification task object //
Created by mlr::makeClassifTask}

\item{model}{mlr WrappedModdel object}

\item{grid.res}{numeric(1)}

\item{x1.lim}{numeric(2)}

\item{x2.lim}{numeric(2)}

\item{colours}{"ESL": colours from the book "Elements of statistical learning"}
}
\value{
ggplot object
}
\description{
Plot decision boundaries, prediciton areas and original data for two features.
}
