% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{factorMapping}
\alias{factorMapping}
\title{Show mapping of factor levels and factor values}
\usage{
factorMapping(x)
}
\arguments{
\item{x}{factor variable}
}
\value{
data.frame
}
\description{
The way R maps levels to values is often not clear so it can help to print out the explicit
mapping.
}
