% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{quantileBins}
\alias{quantileBins}
\title{Put values of vector in bins based on quantiles.}
\usage{
quantileBins(x, probs = c(0, 0.25, 0.5, 0.75, 1),
  include.lowest = TRUE, as.factor = FALSE, values.as.names = FALSE,
  ...)
}
\arguments{
\item{x}{(numeric) Vector you want to bin.}

\item{probs}{(numeric) Default is seq(0, 1, .25) (i.e. quartiles).}

\item{...}{further arguments passed down to the base functions quantile() and cut()}

\item{include.lowest=TRUE}{(logical) Include element with lowest value. Who would not want that
(except the authors of the base cut function)?}

\item{as.factor=FALSE}{(logical) If TRUE returns a factor with quantile boundaries as level
labels. Level labels can be changed via the labels argument.}

\item{values.as.names=FALSE}{(logical) Return named vector with values of the input
so that you can see how bin numbers correspond to values of x.}
}
\value{
integer(length(x)) or factor
}
\description{
Put values of vector in bins based on quantiles.
}
\examples{
quantileBins(1:20, values.as.names = TRUE)

n = 200
x = runif(n, min = 0, max = 1)
x1 = x + rnorm(n, 0, 0.05)
x2 = x + rnorm(n, 0, 0.05)
y = x1 + x2^2 + rnorm(n, 0, 0.1)
plot(x1, x2, cex = quantileBins(y, probs = seq(0,1,.1))/5,
  col = quantileBins(y))
}
